# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "summary": "Returns all products",
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "List of all products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "summary": "Update a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "type": "object",
                "required": [
                    "productId",
                    "supplierId",
                    "name",
                    "price",
                    "sku",
                    "unit"
                ],
                "properties": {
                    "productId": {
                        "type": "integer",
                        "description": "The unique identifier for the product"
                    },
                    "supplierId": {
                        "type": "integer",
                        "description": "The ID of the supplier providing this product"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the product"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the product"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The current price of the product"
                    },
                    "sku": {
                        "type": "string",
                        "description": "Stock keeping unit code for the product"
                    },
                    "unit": {
                        "type": "string",
                        "description": "Unit of measure for the product (e.g., \"piece\")"
                    },
                    "imgName": {
                        "type": "string",
                        "description": "Filename of the product image"
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "description": "Discount percentage (if applicable) expressed as a decimal (e.g., 0.25 for 25%)"
                    }
                }
            },
...
        }
    },
...
}